requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-jwt-extended==4.5.3
flask-bcrypt==1.0.1
werkzeug==2.3.7
python-dotenv==1.0.0
openai==1.3.0
pytesseract==0.3.10
Pillow==10.1.0
PyPDF2==3.0.1
python-docx==0.8.11
pandas==2.1.3
scikit-learn==1.3.2
nltk==3.8.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
requests==2.31.0
